refactor resolvers to its own file
