+++ part8/library-backend/package.json
@@ -21,7 +21,6 @@
     "graphql": "^16.10.0",
     "jsonwebtoken": "^9.0.2",
     "mongoose": "^7.8.6",
-    "mongoose-unique-validator": "^4.0.1",
-    "uuid": "^11.0.5"
+    "mongoose-unique-validator": "^4.0.1"
   }
 }
